+++ calendar-service/src/main/java/com/example/calendarservice/repository/ScheduleRepository.java
@@ -9,10 +9,10 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ScheduleRepository extends JpaRepository<Schedule, Integer> {
+public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
     @Query("SELECT s FROM Schedule s WHERE s.branchId = :branchId AND FUNCTION('MONTH', s.startTime) = :month")
-    List<Schedule> findByBranchIdAndMonth(@Param("branchId") Integer branchId, @Param("month") Integer month);
+    List<Schedule> findByBranchIdAndMonth(@Param("branchId") Long branchId, @Param("month") Integer month);
 
     @Query("SELECT s FROM Schedule s WHERE s.branchId = :branchId AND s.workerId = :userId AND FUNCTION('MONTH', s.startTime) = :month")
-    List<Schedule> findByBranchIdAndMonthAndUserId(@Param("branchId") Integer branchId, @Param("month") Integer month, @Param("userId") Integer userId);
+    List<Schedule> findByBranchIdAndMonthAndUserId(@Param("branchId") Long branchId, @Param("month") Integer month, @Param("userId") Long userId);
 }
